# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "You Ticket API",
        "version": "1.0.11"
    },
    "servers": [
        {
            "url": "http://vercel.com/acttis-moran/acttis-moran-laravel/restapi"
        }
    ],
    "paths": {
        "/matchgame/example": {
            "get": {
                "tags": [
                    "matchgame"
                ],
                "summary": "Devuelve un conjunto de diez Matchgame's aleatorios existentes.",
                "description": "Display a random reduced list of matchgames.",
                "operationId": "edaedc2290b10667220bc316b883daf6",
                "responses": {
                    "200": {
                        "description": "(OK) Se devolvio correctamente el/los matchgames deseados."
                    },
                    "400": {
                        "description": "(Bad Request) Los datos enviados son incorrectos o hay datos obligatorios no enviados"
                    },
                    "404": {
                        "description": "(NotFound) No se encontro informacion"
                    },
                    "500": {
                        "description": "Error en servidor"
                    }
                }
            }
        },
        "/matchgame/matchesby": {
            "get": {
                "tags": [
                    "matchgame"
                ],
                "summary": "Devuelve la información de los matchgame's que depende de los parametros especificados.",
                "description": "Display a listing of matches filter by team, stadium & date.",
                "operationId": "4e1e7ef6ddb6d1cd930381f01f3b41af",
                "parameters": [
                    {
                        "name": "teamId",
                        "in": "query",
                        "description": "Identificador del Equipo que participa en/los Matchgame a obtener.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "stadiumId",
                        "in": "query",
                        "description": "Identificador del Stadium donde se juega el/los Matchgame.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Fecha en la que se realiza el/los Matchgame.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "(OK) Se obtuvo correctamente el/los matchgames deseados."
                    },
                    "400": {
                        "description": "(Bad Request) Los datos enviados son incorrectos o hay datos obligatorios no enviados"
                    },
                    "404": {
                        "description": "(NotFound) No se encontro informacion"
                    },
                    "500": {
                        "description": "Error en servidor"
                    }
                }
            }
        },
        "/matchgame/show/{matchgameId}": {
            "get": {
                "tags": [
                    "matchgame"
                ],
                "summary": "Devuelve la información del matchgame que posee el ID proporcionado.",
                "description": "Display the specified resource.",
                "operationId": "ab36719cd14aabb47b891b6ac5d9d52d",
                "parameters": [
                    {
                        "name": "matchgameId",
                        "in": "path",
                        "description": "Identificador del Matchgame a obtener.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "(OK) Se obtuvo correctamente el matchgame deseado."
                    },
                    "400": {
                        "description": "(Bad Request) Los datos enviados son incorrectos o hay datos obligatorios no enviados"
                    },
                    "404": {
                        "description": "(NotFound) No se encontro informacion"
                    },
                    "500": {
                        "description": "Error en servidor"
                    }
                }
            }
        },
...
    },
...
}